Replace uses of 'Text' for naming with a  'Name' type.
A 'Name' may be unqualified, qualified, or mixfix. An unqualified name
is just text; a qualified name is a sequence of unqualified names,
separated by underscores. An unqualified name may no longer contain
underscores. A mixfix name is a sequence of unqualified names,
interspersed with "holes", spelled "()". Two adjacent unqualified names
are not allowed in a mixfix name, but multiple adjacent holes are
allowed.