Introduce explicit copy and drop calls.
When a variable is used multiple times, uses of the variable are
instrumented with calls to a generic "copy" word. When it's used zero
times, the end of the variable's scope is instrumented with a call to a
generic "drop" word. This introduces a 'declare' keyword for declaring
generic words, and reports non-generic duplicate definitions.