Generate instantiations of generic words.
Rename 'declare' to 'trait' and differentiate 'define' (words) from
'instance' (trait instances).

Unify the inferred type of a definition with its declared type, so that
instantiations within the body are substituted correctly with concrete
types and not type variables.

Add the type of lambda variables to the AST, so that it doesn't need to
be extracted from the inferred type later, e.g., when instrumenting
"copy" and "drop" calls.