Fix subsumption checking for permission labels.
This was a case of swapping two variables.

Previously we were checking that all the labels in the *declared* type
were present in the *inferred* type. So e.g. a function inferred to
require '+Fail +P' but declaring only '+P' would wrongly typecheck,
because all 0 declared types were present in the inferred type.

Now we correctly check that all *inferred* labels are present in the
*declared* type. The declared type may of course contain additional,
unused labels; in the future, these could be reported as warnings.