[wip] Simplify field names.
Rather than a suffix "term.label", treat field access as a regular term
".label".

The current issue is that, if the type environment only contains
signatures, then zonking can't convert them to types because it doesn't
have access to K, and can't because it relies on pure laziness.

The type environment probably just needs to contain actual types, and
maybe zonking can collect instantiations of types as it goes. I don't
think it'd work if we did it after zonking, because the final inferred
type needs to have no unresolved field accesses in order to be checked
against the signature.