Improve pretty-printing of types.
This adds an unqualified name to type variables as a "hint" for how to
display them. The hint of a user-specified type variable is taken
directly from the source code. When the compiler generates a fresh type
variable during inference, it tries to preserve a user-specified hint;
if it can't, it tries to generate a descriptive name, such as 'Capture1'
for the first variable captured by a closure or 'VarX' for the type of a
local variable named 'x'.

When multiple type variables share the same name hint, during pretty
printing they are suffixed with the index ("_1", "_2", ...)  of their
first (depth-first) occurrence in the whole pretty-printed type.

This currently suffers from the limitation that variables are *not*
disambiguated when pretty-printing multiple types, since they don't
share the same context. And the output could be further improved by
omitting implicit stack and permission variables as a user would do.

Before:

    //type \(+ 1)
    <T202262..., +T202274, +T202270> ((T202262... -> T202262..., <T202267...> ((T202267..., _::Int32 -> T202267..., _::Int32 +T202270)) +T202274))

    //type ask dup say fail
    <T203082..., +T203080, T203075...> ((T203075..., _::List<_::Char> -> T203082... +_::IO +_::Fail +T203080))

After (~35% reduction):

    //type \(+ 1)
    <S..., +P_1, +P_2> ((S... -> S..., <R...> ((R..., _::Int32 -> R..., _::Int32 +P_2)) +P_1))

    //type ask dup say fail
    <S..., +P, R...> ((R..., _::List<_::Char> -> S... +_::IO +_::Fail +P))